+++ src/main/java/step4/LottoBucket.java
@@ -1,4 +1,4 @@
-package step3;
+package step4;
 
 import java.util.ArrayList;
 import java.util.Iterator;
