+++ src/main/java/step4/LottoEntry.java
@@ -1,4 +1,4 @@
-package step3;
+package step4;
 
 import java.util.LinkedHashSet;
 import java.util.List;
