+++ src/main/java/step4/LottoPrice.java
@@ -1,4 +1,4 @@
-package step3;
+package step4;
 
 public class LottoPrice {
     private final int unitPrice;
