+++ src/main/java/step4/enums/DrawResult.java
@@ -1,4 +1,4 @@
-package step3.enums;
+package step4.enums;
 
 import java.util.Optional;
 import java.util.stream.Stream;
