+++ src/main/java/step4/util/LottoManualNumberGenerator.java
@@ -1,4 +1,4 @@
-package step3.util;
+package step4.util;
 
 import java.util.Iterator;
 import java.util.List;
