+++ src/main/java/step4/util/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package step3.util;
+package step4.util;
 
 import java.util.List;
