+++ src/main/java/step4/view/LottoInputView.java
@@ -1,4 +1,4 @@
-package step3.view;
+package step4.view;
 
 import java.util.Arrays;
 import java.util.List;
