+++ src/test/java/step4/DrawResultTest.java
@@ -1,10 +1,8 @@
-package step3;
+package step4;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.enums.DrawResult;
-
-import java.util.Arrays;
+import step4.enums.DrawResult;
 
 import static org.assertj.core.api.Assertions.assertThat;
