+++ src/test/java/step4/DrawnLottoTest.java
@@ -1,8 +1,8 @@
-package step3;
+package step4;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.enums.DrawResult;
+import step4.enums.DrawResult;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
