+++ src/test/java/step4/LottoBucketTest.java
@@ -1,4 +1,4 @@
-package step3;
+package step4;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
